moved pages into their own folder and created style files for each one
