+++ client/src/components/Form/styles.css
@@ -40,7 +40,7 @@
 }
 
 .Form-submit {
-    margin-top: 94px;
+    margin-top: 92px;
     background: none;
     outline: none;
     border: none;
