+++ client/src/components/index.js
@@ -1,2 +1,3 @@
 export { default as Header } from "./Header";
 export { default as Layout } from "./Layout";
+export { default as Form } from "./Form"
