--- client/src/pages/RegisterEmp.jsx
@@ -1,11 +1,0 @@
-import Form from "../components/Form"
-
-const RegisterEmp = () => {
-  return (
-    <div>
-      <Form />
-    </div>
-  )
-}
-
-export default RegisterEmp
