+++ client/src/pages/RegisterEmp/index.jsx
@@ -1,0 +1,13 @@
+import { Header, Form } from "../../components";
+import "./styles.css"
+
+const RegisterEmp = () => {
+  return (
+    <div>
+      <Header title={"Register new employee"} />
+      <Form />
+    </div>
+  );
+};
+
+export default RegisterEmp;
