+++ client/src/pages/index.js
@@ -1,6 +1,6 @@
 export { default as ActiveEmps } from "./ActiveEmps";
 export { default as EditEmp } from "./EditEmp";
 export { default as EmpDetails } from "./EmpDetails";
-export { default as Logs } from "./Logs";
+export { default as EmpLogs } from "./EmpLogs";
 export { default as RegisterEmp } from "./RegisterEmp";
 export { default as WorkedHours } from "./WorkedHours";
